Split interrupts testing into its own file
